Refactor Legend class into new annotations.py module
