+++ bokeh/models/__init__.py
@@ -3,6 +3,7 @@ from __future__ import absolute_import
 # This file is excluded from flake8 checking in setup.cfg
 
 from .actions import *
+from .annotations import *
 from .axes import *
 from .formatters import *
 from .glyphs import *
