+++ bokeh/models/renderers.py
@@ -5,9 +5,8 @@ types that Bokeh supports.
 from __future__ import absolute_import
 
 from ..plot_object import PlotObject
-from ..properties import Int, String, Enum, Instance, List, Dict, Tuple, Include
-from ..mixins import LineProps, TextProps, FillProps
-from ..enums import Units, Orientation, RenderLevel
+from ..properties import String, Enum, Instance
+from ..enums import Units, RenderLevel
 from ..validation.errors import BAD_COLUMN_NAME, MISSING_GLYPH, NO_SOURCE_FOR_GLYPH
 from .. import validation
 
@@ -82,73 +81,6 @@ class GlyphRenderer(Renderer):
     level = Enum(RenderLevel, default="glyph", help="""
     Specifies the level in which to render the glyph.
     """)
-
-# TODO: (bev) This should really go in a separate module
-class Legend(Renderer):
-    """ Render informational legends for a plot.
-
-    """
-
-    plot = Instance(".models.plots.Plot", help="""
-    The Plot to which this Legend is attached.
-    """)
-
-    orientation = Enum(Orientation, help="""
-    The location where the legend should draw itself.
-    """)
-
-    border_props = Include(LineProps, help="""
-    The %s for the legend border outline.
-    """)
-
-    background_props = Include(FillProps, help="""
-    The %s for the legend background style.
-    """)
-
-    label_props = Include(TextProps, help="""
-    The %s for the legend labels.
-    """)
-
-    label_standoff = Int(15, help="""
-    The distance (in pixels) to separate the label from its associated glyph.
-    """)
-
-    label_height = Int(20, help="""
-    The height (in pixels) of the area that legend labels should occupy.
-    """)
-
-    label_width = Int(50, help="""
-    The width (in pixels) of the area that legend labels should occupy.
-    """)
-
-    glyph_height = Int(20, help="""
-    The height (in pixels) that the rendered legend glyph should occupy.
-    """)
-
-    glyph_width = Int(20, help="""
-    The width (in pixels) that the rendered legend glyph should occupy.
-    """)
-
-    legend_padding = Int(10, help="""
-    Amount of padding around the legend.
-    """)
-
-    legend_spacing = Int(3, help="""
-    Amount of spacing between legend entries.
-    """)
-
-    legends = List(Tuple(String, List(Instance(GlyphRenderer))), help="""
-    A list of tuples that maps text labels to the legend to corresponding
-    renderers that should draw sample representations for those labels.
-
-    .. note::
-        The ``legends`` attribute may also be set from a dict or OrderedDict,
-        but note that if a dict is used, the order of the legend entries is
-        unspecified.
-
-    """).accepts(
-        Dict(String, List(Instance(GlyphRenderer))), lambda d: list(d.items())
-    )
 
 class GuideRenderer(Renderer):
     """ A base class for all guide renderer types. ``GuideRenderer`` is
